refactor(test): moved tests to a unit test folder
